Update deletes with the sql function

+++ routes/securityAnswer/deleteSecurityAnswer.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import { idSchema } from '../../types';
 import { HttpError } from '../../middleware/errorMiddleware';
-import pool from '../../db';
+import { sql } from '../../db';
 import authMiddleware from '../../middleware/auth';
 export default (router: Router) => {
   router.delete('/:id', authMiddleware(), async (req, res) => {
@@ -10,17 +10,10 @@ export default (router: Router) => {
       throw new HttpError(422, 'Invalid ID');
     }
     const securityAnswerId = validationResult.data;
-
-    const deleteQuery = `
-      DELETE FROM security_answers 
-      WHERE id = $1
-      RETURNING *`;
-    const deleteValues = [securityAnswerId];
-
-    const deleteResult = await pool.query(deleteQuery, deleteValues);
-    if (deleteResult.rows.length === 0) {
-      throw new HttpError(404, 'Security answer not found');
-    }
-    res.json({ message: 'Security answer deleted successfully' });
+    const userId = req.user!.id;
+    const query = 'DELETE FROM security_answers  WHERE id = :securityAnswerId AND user_id = :userId';
+    const SQL_DELETE_SAVING = sql<{ id: string; user_id: string }, Record<string, never>>(query);
+    await SQL_DELETE_SAVING({id:securityAnswerId, user_id: userId }).exec();
+    return res.json({ message: 'Answer deleted successfully' });
   });
 };
